various fixes
- As we updated the menu item "About Scandic Fusion" last time, it is
  now clickable and redirects to "Contact Us" page. It is good, just I
think it will make sense to pull "Contact Us" on top of the dropdown
menu (make it the 1st item), since if it is a destination for the parent
page, it is logical to see the page itself as close as possible.

Now respects the order you set in the CMS.

-  CMS exception error in: "Tech stack" under Settings, "Jobs" under
   Settings. I haven't noticed it in any other places, so I think it is
safe to fix it now.

Fixed. (removed tab)